Create `validate-base` plugin
