+++ src/plugins/validate-semantic/validators/paths.js
