+++ src/plugins/validate-semantic/validators/security.js
